Use new reflection API in tests
